requirements.txt
# Core Data Science Stack
pandas>=2.0.0
numpy>=1.24.0

# Google Trends API
pytrends>=4.9.0

# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Browser Automation (for dynamic content)
selenium>=4.15.0
playwright>=1.40.0

# User-Agent Rotation (Anti-Bot)
fake-useragent>=1.4.0

# Data Visualization (for notebooks)
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration Management
pyyaml>=6.0

# Logging
colorlog>=6.7.0

# Rate Limiting
ratelimit>=2.2.1

# HTTP Requests Enhancement
requests-html>=0.10.0

# Optional: Progress bars
tqdm>=4.66.0
